Downloads/image-project-mknows-main/image-project-mknows-main/requirements.txt
protobuf==3.20.0
django==4.2.4
djangorestframework==3.14.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
keras==2.13.1
tensorflow==2.13
opencv-python == 4.8.1.78
deepface == 0.0.79
tensorflow==2.13.0
Pillow==10.0.0
paddleocr==2.7.0.3
paddlepaddle
